+++ pkg/compactor/compactor_test.go
@@ -6,6 +6,9 @@ import (
 	"encoding/json"
 	"errors"
 	"flag"
+	"fmt"
+	"io/ioutil"
+	"os"
 	"strings"
 	"testing"
 	"time"
@@ -20,6 +23,8 @@ import (
 	"github.com/stretchr/testify/require"
 	"gopkg.in/yaml.v2"
 
+	"github.com/cortexproject/cortex/pkg/ring"
+	"github.com/cortexproject/cortex/pkg/ring/kv/consul"
 	cortex_tsdb "github.com/cortexproject/cortex/pkg/storage/tsdb"
 	"github.com/cortexproject/cortex/pkg/util/flagext"
 	cortex_testutil "github.com/cortexproject/cortex/pkg/util/test"
@@ -74,14 +79,16 @@ func TestCompactor_ShouldDoNothingOnNoUserBlocks(t *testing.T) {
 	bucketClient := &cortex_tsdb.BucketClientMock{}
 	bucketClient.MockIter("", []string{}, nil)
 
-	c, _, logs, registry := prepare(t, bucketClient)
+	c, _, logs, registry, cleanup := prepare(t, prepareConfig(), bucketClient)
+	defer cleanup()
+	c.Start()
 
 	// Wait until a run has completed.
 	cortex_testutil.Poll(t, time.Second, 1.0, func() interface{} {
 		return prom_testutil.ToFloat64(c.compactionRunsCompleted)
 	})
 
-	c.Shutdown()
+	c.Stop()
 
 	assert.Equal(t, []string{
 		`level=info msg="discovering users from bucket"`,
@@ -162,14 +169,16 @@ func TestCompactor_ShouldRetryOnFailureWhileDiscoveringUsersFromBucket(t *testin
 	bucketClient := &cortex_tsdb.BucketClientMock{}
 	bucketClient.MockIter("", nil, errors.New("failed to iterate the bucket"))
 
-	c, _, logs, registry := prepare(t, bucketClient)
+	c, _, logs, registry, cleanup := prepare(t, prepareConfig(), bucketClient)
+	defer cleanup()
+	c.Start()
 
 	// Wait until all retry attempts have completed.
 	cortex_testutil.Poll(t, time.Second, 1.0, func() interface{} {
 		return prom_testutil.ToFloat64(c.compactionRunsFailed)
 	})
 
-	c.Shutdown()
+	c.Stop()
 
 	// Ensure the bucket iteration has been retried the configured number of times.
 	bucketClient.AssertNumberOfCalls(t, "Iter", 3)
@@ -261,7 +270,9 @@ func TestCompactor_ShouldIterateOverUsersAndRunCompaction(t *testing.T) {
 	bucketClient.MockGet("user-1/01DTVP434PA9VFXSW2JKB3392D/meta.json", mockBlockMetaJSON("01DTVP434PA9VFXSW2JKB3392D"), nil)
 	bucketClient.MockGet("user-2/01DTW0ZCPDDNV4BV83Q2SV4QAZ/meta.json", mockBlockMetaJSON("01DTW0ZCPDDNV4BV83Q2SV4QAZ"), nil)
 
-	c, tsdbCompactor, logs, registry := prepare(t, bucketClient)
+	c, tsdbCompactor, logs, registry, cleanup := prepare(t, prepareConfig(), bucketClient)
+	defer cleanup()
+	c.Start()
 
 	// Mock the compactor as if there's no compaction to do,
 	// in order to simplify tests (all in all, we just want to
@@ -274,7 +285,7 @@ func TestCompactor_ShouldIterateOverUsersAndRunCompaction(t *testing.T) {
 		return prom_testutil.ToFloat64(c.compactionRunsCompleted)
 	})
 
-	c.Shutdown()
+	c.Stop()
 
 	// Ensure a plan has been executed for the blocks of each user.
 	tsdbCompactor.AssertNumberOfCalls(t, "Plan", 2)
@@ -314,6 +325,176 @@ func TestCompactor_ShouldIterateOverUsersAndRunCompaction(t *testing.T) {
 	`), testedMetrics...))
 }
 
+func TestCompactor_ShouldCompactAllUsersOnShardingEnabledButOnlyOneInstanceRunning(t *testing.T) {
+	t.Parallel()
+
+	// Mock the bucket to contain two users, each one with one block.
+	bucketClient := &cortex_tsdb.BucketClientMock{}
+	bucketClient.MockIter("", []string{"user-1", "user-2"}, nil)
+	bucketClient.MockIter("user-1/", []string{"user-1/01DTVP434PA9VFXSW2JKB3392D"}, nil)
+	bucketClient.MockIter("user-2/", []string{"user-2/01DTW0ZCPDDNV4BV83Q2SV4QAZ"}, nil)
+	bucketClient.MockGet("user-1/01DTVP434PA9VFXSW2JKB3392D/meta.json", mockBlockMetaJSON("01DTVP434PA9VFXSW2JKB3392D"), nil)
+	bucketClient.MockGet("user-2/01DTW0ZCPDDNV4BV83Q2SV4QAZ/meta.json", mockBlockMetaJSON("01DTW0ZCPDDNV4BV83Q2SV4QAZ"), nil)
+
+	cfg := prepareConfig()
+	cfg.ShardingEnabled = true
+	cfg.ShardingRing.InstanceID = "compactor-1"
+	cfg.ShardingRing.KVStore.Mock = consul.NewInMemoryClient(ring.GetCodec())
+
+	c, tsdbCompactor, logs, _, cleanup := prepare(t, cfg, bucketClient)
+	defer cleanup()
+	c.Start()
+
+	// Mock the compactor as if there's no compaction to do,
+	// in order to simplify tests (all in all, we just want to
+	// test our logic and not TSDB compactor which we expect to
+	// be already tested).
+	tsdbCompactor.On("Plan", mock.Anything).Return([]string{}, nil)
+
+	// Wait until a run has completed.
+	cortex_testutil.Poll(t, 5*time.Second, 1.0, func() interface{} {
+		return prom_testutil.ToFloat64(c.compactionRunsCompleted)
+	})
+
+	c.Stop()
+
+	// Ensure a plan has been executed for the blocks of each user.
+	tsdbCompactor.AssertNumberOfCalls(t, "Plan", 2)
+
+	assert.Equal(t, []string{
+		`level=info msg="waiting until compactor is ACTIVE in the ring"`,
+		`level=info msg="compactor is ACTIVE in the ring"`,
+		`level=info msg="discovering users from bucket"`,
+		`level=info msg="discovered users from bucket" users=2`,
+		`level=info msg="starting compaction of user blocks" user=user-1`,
+		`level=info msg="start sync of metas"`,
+		`level=info msg="start of GC"`,
+		`level=info msg="start of compaction"`,
+		`level=info msg="compaction iterations done"`,
+		`level=info msg="successfully compacted user blocks" user=user-1`,
+		`level=info msg="starting compaction of user blocks" user=user-2`,
+		`level=info msg="start sync of metas"`,
+		`level=info msg="start of GC"`,
+		`level=info msg="start of compaction"`,
+		`level=info msg="compaction iterations done"`,
+		`level=info msg="successfully compacted user blocks" user=user-2`,
+	}, removeMetaFetcherLogs(strings.Split(strings.TrimSpace(logs.String()), "\n")))
+}
+
+func TestCompactor_ShouldCompactOnlyUsersOwnedByTheInstanceOnShardingEnabledAndMultipleInstancesRunning(t *testing.T) {
+	t.Parallel()
+
+	numUsers := 100
+
+	// Setup user IDs
+	userIDs := make([]string, 0, numUsers)
+	for i := 1; i <= numUsers; i++ {
+		userIDs = append(userIDs, fmt.Sprintf("user-%d", i))
+	}
+
+	// Mock the bucket to contain all users, each one with one block.
+	bucketClient := &cortex_tsdb.BucketClientMock{}
+	bucketClient.MockIter("", userIDs, nil)
+	for _, userID := range userIDs {
+		bucketClient.MockIter(userID+"/", []string{"user-1/01DTVP434PA9VFXSW2JKB3392D"}, nil)
+		bucketClient.MockGet(userID+"/01DTVP434PA9VFXSW2JKB3392D/meta.json", mockBlockMetaJSON("01DTVP434PA9VFXSW2JKB3392D"), nil)
+	}
+
+	// Create a shared KV Store
+	kvstore := consul.NewInMemoryClient(ring.GetCodec())
+
+	// Create two compactors
+	compactors := []*Compactor{}
+	logs := []*bytes.Buffer{}
+
+	for i := 1; i <= 2; i++ {
+		cfg := prepareConfig()
+		cfg.ShardingEnabled = true
+		cfg.ShardingRing.InstanceID = fmt.Sprintf("compactor-%d", i)
+		cfg.ShardingRing.InstanceAddr = fmt.Sprintf("127.0.0.%d", i)
+		cfg.ShardingRing.KVStore.Mock = kvstore
+
+		c, tsdbCompactor, l, _, cleanup := prepare(t, cfg, bucketClient)
+		defer c.Stop()
+		defer cleanup()
+
+		compactors = append(compactors, c)
+		logs = append(logs, l)
+
+		// Mock the compactor as if there's no compaction to do,
+		// in order to simplify tests (all in all, we just want to
+		// test our logic and not TSDB compactor which we expect to
+		// be already tested).
+		tsdbCompactor.On("Plan", mock.Anything).Return([]string{}, nil)
+	}
+
+	// Wait until each compactor sees all ACTIVE compactors in the ring
+	for _, c := range compactors {
+		cortex_testutil.Poll(t, 10*time.Second, len(compactors), func() interface{} {
+			rs, err := c.ring.GetAll()
+			if err != nil {
+				return 0
+			}
+
+			numActive := 0
+			for _, i := range rs.Ingesters {
+				if i.GetState() == ring.ACTIVE {
+					numActive++
+				}
+			}
+
+			return numActive
+		})
+	}
+
+	// Start all compactors
+	for _, c := range compactors {
+		c.Start()
+	}
+
+	// Wait until a run has been completed on each compactor
+	for _, c := range compactors {
+		cortex_testutil.Poll(t, 10*time.Second, 1.0, func() interface{} {
+			return prom_testutil.ToFloat64(c.compactionRunsCompleted)
+		})
+	}
+
+	// Ensure that each user has been compacted by the correct instance
+	for _, userID := range userIDs {
+		_, l, err := findCompactorByUserID(compactors, logs, userID)
+		require.NoError(t, err)
+		assert.Contains(t, l.String(), fmt.Sprintf(`level=info msg="successfully compacted user blocks" user=%s`, userID))
+	}
+}
+
+func findCompactorByUserID(compactors []*Compactor, logs []*bytes.Buffer, userID string) (*Compactor, *bytes.Buffer, error) {
+	var compactor *Compactor
+	var log *bytes.Buffer
+
+	for i, c := range compactors {
+		owned, err := c.ownUser(userID)
+		if err != nil {
+			return nil, nil, err
+		}
+
+		// Ensure the user is not owned by multiple compactors
+		if owned && compactor != nil {
+			return nil, nil, fmt.Errorf("user %s owned by multiple compactors", userID)
+		}
+		if owned {
+			compactor = c
+			log = logs[i]
+		}
+	}
+
+	// Return an error if we've not been able to find it
+	if compactor == nil {
+		return nil, nil, fmt.Errorf("user %s not owned by any compactor", userID)
+	}
+
+	return compactor, log, nil
+}
+
 func removeMetaFetcherLogs(input []string) []string {
 	out := make([]string, 0, len(input))
 
@@ -326,12 +507,28 @@ func removeMetaFetcherLogs(input []string) []string {
 	return out
 }
 
-func prepare(t *testing.T, bucketClient *cortex_tsdb.BucketClientMock) (*Compactor, *tsdbCompactorMock, *bytes.Buffer, prometheus.Gatherer) {
+func prepareConfig() Config {
 	compactorCfg := Config{}
+	flagext.DefaultValues(&compactorCfg)
+
+	return compactorCfg
+}
+
+func prepare(t *testing.T, compactorCfg Config, bucketClient *cortex_tsdb.BucketClientMock) (*Compactor, *tsdbCompactorMock, *bytes.Buffer, prometheus.Gatherer, func()) {
 	storageCfg := cortex_tsdb.Config{}
-	flagext.DefaultValues(&compactorCfg, &storageCfg)
+	flagext.DefaultValues(&storageCfg)
+
 	compactorCfg.retryMinBackoff = 0
 	compactorCfg.retryMaxBackoff = 0
+
+	// Create a temporary directory for compactor data.
+	dataDir, err := ioutil.TempDir(os.TempDir(), "compactor-test")
+	require.NoError(t, err)
+
+	compactorCfg.DataDir = dataDir
+	cleanup := func() {
+		os.RemoveAll(dataDir)
+	}
 
 	tsdbCompactor := &tsdbCompactorMock{}
 	logs := &bytes.Buffer{}
@@ -342,7 +539,7 @@ func prepare(t *testing.T, bucketClient *cortex_tsdb.BucketClientMock) (*Compact
 	c, err := newCompactor(ctx, cancelCtx, compactorCfg, storageCfg, bucketClient, tsdbCompactor, logger, registry)
 	require.NoError(t, err)
 
-	return c, tsdbCompactor, logs, registry
+	return c, tsdbCompactor, logs, registry, cleanup
 }
 
 type tsdbCompactorMock struct {